Changed check condition when a user types !use <Something> | Rogue
A mongodb .UpdateOne function (being used here), will spit out information on whether or not it has modified any files. If it spits out a 0 (no files modified), then we know the user isn't registered. Changed check from simply finding a result to finding a result that was changed.

if (res) always passes, if (res.n>0) only passes if n (the number of modified files) has changed a file