Refactored how !deckstats displays information
-Changed deckstats from... looping through query output array (all the information about decks) and adding new fields in the embed for each

to

looping through query output array, adding each deckname and win/loss information to strings, then printing out that one big string in a single embed field.

The result is a much more compressed and clean way to display this data.

Also, I have set up checks for when the embed gets too large. If a server has a million decks played across a season, they will hit the 1024 character limit. Bot will only display the decks with the most games played in the before it hits the cut off